Add orders endpoints

+++ src/products/products.controller.ts
@@ -1,6 +1,8 @@
-import { Controller, Get, Delete, Param, Post, Body } from '@nestjs/common';
+import { Controller, Get, Delete, Param, Post, Body, Put } from '@nestjs/common';
 import { ProductsService } from './products.service';
 import { CreateProductDTO } from './dtos/create-product.dto';
+import { ParseUUIDPipe, NotFoundException } from '@nestjs/common';
+import { UpdateProductDTO } from './dtos/update-product.dto';
 
 @Controller('products')
 export class ProductsController {
@@ -11,17 +13,33 @@ export class ProductsController {
         return this.productsService.getAll();
     }
     @Get('/:id')
-    getById(@Param('id') id: string) {
-        return this.productsService.getById(id)
+    getById(@Param('id', new ParseUUIDPipe()) id: string) {
+        const prod = this.productsService.getById(id);
+        if (!prod) throw new NotFoundException('Product not found');
+        return prod;
     }
     @Delete('/:id')
-    deleteById(@Param('id') id: string) {
+    deleteById(@Param('id', new ParseUUIDPipe()) id: string) {
+        if (!this.productsService.getById(id)) {
+            throw new NotFoundException('Product not found');
+        }
         this.productsService.deleteById(id);
         return { success: true };
     }
     @Post('/')
     create(@Body() productData: CreateProductDTO) {
         return this.productsService.create(productData);
+    }
+    @Put('/:id')
+    update(
+        @Param('id', new ParseUUIDPipe()) id: string,
+        @Body() productData: UpdateProductDTO,
+    ) {
+        if (!this.productsService.getById(id)) { }
+        throw new NotFoundException('Product not found');
 
+
+        this.productsService.updateById(id, productData);
+        return { success: true };
     }
 }
